chore: move test from search-view to search-message-panel
