+++ src/views/search/panel/message/tests/search-message-panel.test.tsx
@@ -7,10 +7,11 @@
 import React from 'react';
 
 import { act, waitFor } from '@testing-library/react';
+import { FOLDERS } from '@zextras/carbonio-ui-commons';
 import { NavigateFunction, useParams } from 'react-router-dom';
 import type { Mock } from 'vitest';
 
-import { setupTest, screen } from '@test-setup';
+import { setupTest, screen, within } from '@test-setup';
 import { generateMessage, populateMessagesInEmailStore } from '__test__/generators/generateMessage';
 import { API_REQUEST_STATUS } from 'constants/index';
 import { setSearchResultsByMessage, updateMessageStatus } from 'store/emails/store';
@@ -59,5 +60,59 @@ describe('Message Panel', () => {
 		});
 
 		expect(mockNavigateSpy).toHaveBeenCalledWith('/search', { replace: true });
+	});
+
+	it('should open warning dialog when clicking edit action on scheduled draft message in preview panel', async () => {
+		(useParams as Mock).mockReturnValue({ messageId: '1' });
+		setSearchResultsByMessage(
+			[
+				generateMessage({
+					id: '1',
+					isComplete: true,
+					subject: 'Scheduled Draft Subject',
+					body: 'Test Message body',
+					isScheduled: true,
+					folderId: FOLDERS.DRAFTS,
+					isDraft: true
+				})
+			],
+			false
+		);
+
+		await waitFor(() => {
+			updateMessageStatus('1', API_REQUEST_STATUS.fulfilled);
+		});
+
+		const { user } = setupTest(<SearchMessagePanel messageId="1" />);
+
+		// Wait for the message panel to appear
+		expect(await screen.findByTestId(`SearchMessagePanel-1`)).toBeInTheDocument();
+
+		// Wait for the message content to load
+		await waitFor(() => {
+			expect(screen.getByText('Scheduled Draft Subject')).toBeInTheDocument();
+		});
+
+		// Find and click the edit action button in the preview header
+		const editButton = await screen.findByTestId('icon: Edit2Outline');
+		expect(editButton).toBeInTheDocument();
+
+		await user.click(editButton);
+
+		// Verify warning modal appears
+		const modal = await screen.findByTestId('modal');
+		expect(modal).toBeInTheDocument();
+
+		// Verify modal title
+		expect(within(modal).getByText('label.warning')).toBeInTheDocument();
+
+		// Verify modal message about delayed sending
+		expect(within(modal).getByText('messages.edit_schedule_warning')).toBeInTheDocument();
+
+		// Verify "Edit anyway" button exists
+		const editAnywayButton = within(modal).getByRole('button', {
+			name: 'action.edit_anyway'
+		});
+		expect(editAnywayButton).toBeInTheDocument();
 	});
 });
