+++ src/views/search/test/search-view.test.tsx
@@ -1110,73 +1110,6 @@ describe.skip('SearchView', () => {
 				expect(totalItemsSelectedAfterAction).toHaveLength(1);
 			});
 		});
-
-		it('should open warning dialog when clicking edit action on scheduled draft message in preview panel', async () => {
-			const { queryChip } = setupSearchViewTest({ viewBy: 'message', query: 'hello' });
-
-			const scheduledTime = Date.now() + 3600000; // 1 hour in the future
-			const messageId = '100';
-			const scheduledDraftMessage = getSoapMessage(messageId, {
-				su: 'Scheduled Draft Subject',
-				f: 'd', // Draft flag
-				l: FOLDERS.DRAFTS,
-				autoSendTime: scheduledTime
-			});
-
-			createSoapAPIInterceptor<SearchRequest, SearchResponse>('Search', {
-				m: [scheduledDraftMessage],
-				more: false
-			});
-
-			createSoapAPIInterceptor<GetMsgRequest, GetMsgResponse>('GetMsg', {
-				m: [scheduledDraftMessage]
-			});
-
-			const mockUseQuery = vi.fn();
-			mockUseQuery.mockReturnValue([[queryChip], noop]);
-			const resultsHeader = (props: { label: string }): ReactElement => <>{props.label}</>;
-			const searchViewProps: SearchViewProps = {
-				useQuery: mockUseQuery,
-				ResultsHeader: resultsHeader,
-				useDisableSearch: () => [false, noop]
-			};
-
-			const { user } = setupTest(<SearchView {...searchViewProps} />, {
-				initialEntries: [`/message/${messageId}`]
-			});
-
-			// Wait for the message panel to appear
-			expect(await screen.findByTestId(`SearchMessagePanel-${messageId}`)).toBeInTheDocument();
-
-			// Wait for the message content to load
-			await waitFor(() => {
-				expect(screen.getByText('Scheduled Draft Subject')).toBeInTheDocument();
-			});
-
-			// Find and click the edit action button in the preview header
-			const editButton = await screen.findByTestId('icon: Edit2Outline');
-			expect(editButton).toBeInTheDocument();
-
-			await act(async () => {
-				await user.click(editButton);
-			});
-
-			// Verify warning modal appears
-			const modal = await screen.findByTestId('modal');
-			expect(modal).toBeInTheDocument();
-
-			// Verify modal title
-			expect(within(modal).getByText('label.warning')).toBeInTheDocument();
-
-			// Verify modal message about delayed sending
-			expect(within(modal).getByText('messages.edit_schedule_warning')).toBeInTheDocument();
-
-			// Verify "Edit anyway" button exists
-			const editAnywayButton = within(modal).getByRole('button', {
-				name: 'action.edit_anyway'
-			});
-			expect(editAnywayButton).toBeInTheDocument();
-		});
 	});
 
 	it('should not call search API if query empty', async () => {
